# core-lending-collateral-management-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-collateral-management
  description: Collateral & Guarantee Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/general-ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/guarantees/{guaranteeRecordId}:
    get:
      tags:
        - GuaranteeRecord
      summary: Get a GuaranteeRecord by ID
      operationId: getById
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeRecordDTO'
    put:
      tags:
        - GuaranteeRecord
      summary: Update a GuaranteeRecord
      operationId: update
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeRecordDTO'
    delete:
      tags:
        - GuaranteeRecord
      summary: Delete a GuaranteeRecord
      operationId: delete
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/guarantees/{guaranteeRecordId}/parties/{guaranteePartyId}:
    get:
      tags:
        - GuaranteeParty
      summary: Get a Guarantee Party by ID
      operationId: getById_2
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteePartyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteePartyDTO'
    put:
      tags:
        - GuaranteeParty
      summary: Update a Guarantee Party
      operationId: update_2
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteePartyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteePartyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteePartyDTO'
    delete:
      tags:
        - GuaranteeParty
      summary: Delete a Guarantee Party record
      operationId: delete_2
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteePartyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/guarantees:
    get:
      tags:
        - GuaranteeRecord
      summary: List or Search Guarantee Records
      operationId: findAll
      parameters:
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestGuaranteeRecordDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseGuaranteeRecordDTO'
    post:
      tags:
        - GuaranteeRecord
      summary: Create a GuaranteeRecord
      operationId: create
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeRecordDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeRecordDTO'
...
  /api/v1/guarantees/{guaranteeRecordId}/parties:
    get:
      tags:
        - GuaranteeParty
      summary: List or Search Guarantee Parties
      operationId: findAll_2
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestGuaranteePartyDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseGuaranteePartyDTO'
    post:
      tags:
        - GuaranteeParty
      summary: Create a Guarantee Party record
      operationId: create_2
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteePartyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteePartyDTO'
...
components:
  schemas:
    GuaranteeRecordDTO:
      required:
        - guaranteeAmount
        - guaranteeStatus
        - guaranteeType
        - loanApplicationId
        - startDate
        - termsConditions
      type: object
      properties:
        guaranteeRecordId:
          type: string
          format: uuid
          readOnly: true
        loanApplicationId:
          type: string
          format: uuid
        loanServicingCaseId:
          type: string
          format: uuid
        guaranteeType:
          type: string
          enum:
            - PERSONAL
            - CORPORATE
            - STANDBY_LETTER
            - BANK_GUARANTEE
        guaranteeAmount:
          minimum: 0
          exclusiveMinimum: true
          type: number
        guaranteeStatus:
          type: string
          enum:
            - ACTIVE
            - EXPIRED
            - TERMINATED
            - PENDING
        termsConditions:
          maxLength: 2000
          minLength: 1
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GuaranteePartyDTO:
      required:
        - guaranteeRecordId
        - guaranteeRole
        - partyId
      type: object
      properties:
        guaranteePartyId:
          type: string
          format: uuid
          readOnly: true
        guaranteeRecordId:
          type: string
          format: uuid
        partyId:
          type: string
          format: uuid
        guaranteeRole:
          type: string
          enum:
            - GUARANTOR
            - BENEFICIARY
            - AGENT
        additionalDetails:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestGuaranteeRecordDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/GuaranteeRecordDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponseGuaranteeRecordDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/GuaranteeRecordDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestGuaranteePartyDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/GuaranteePartyDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseGuaranteePartyDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/GuaranteePartyDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...